Renamed to chatcodes - thanks Jitterted
