+++ build.gradle
@@ -4,8 +4,8 @@ plugins {
     id 'org.jetbrains.intellij' version '0.4.7'
 }
 
-group 'com.jitterted'
-version '0.1.0'
+group 'dev.chatcodes'
+version '0.0.1'
 
 // make sure this is at most 1.8 to be compatible with IntelliJ's JRE
 sourceCompatibility = 1.8
