+++ settings.gradle
@@ -1,2 +1,2 @@
-rootProject.name = 'jitterchat'
+rootProject.name = 'ChatCodes'
