+++ src/main/java/dev/chatcodes/ChatCodesConfigurable.java
@@ -1,0 +1,40 @@
+package dev.chatcodes;
+
+import com.intellij.credentialStore.CredentialAttributes;
+import com.intellij.credentialStore.CredentialAttributesKt;
+import com.intellij.credentialStore.Credentials;
+import com.intellij.ide.passwordSafe.PasswordSafe;
+import com.intellij.openapi.options.Configurable;
+import com.intellij.openapi.options.ConfigurationException;
+
+import javax.swing.*;
+
+public class ChatCodesConfigurable implements Configurable {
+
+    private ConfigForm settingsPane;
+
+    public JComponent createComponent() {
+        settingsPane = new ConfigForm();
+        return settingsPane.getSettingsPanel();
+    }
+
+    @Override
+    public boolean isModified() {
+        return true;
+    }
+
+    @Override
+    public void apply() throws ConfigurationException {
+        CredentialAttributes credentialAttributes = new CredentialAttributes(
+            CredentialAttributesKt.generateServiceName("ChatCodes", "TwitchOAuthToken")
+        );
+
+        Credentials credentials = new Credentials("", settingsPane.getOAuthTokenValue());
+        PasswordSafe.getInstance().set(credentialAttributes, credentials);
+    }
+
+    @Override
+    public String getDisplayName() {
+        return "ChatCodes";
+    }
+}
