+++ src/main/java/dev/chatcodes/ChatToolWindow.form
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="com.jitterted.ChatToolWindow">
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="dev.chatcodes.ChatToolWindow">
   <grid id="27dc6" binding="chatContentPanel" layout-manager="GridLayoutManager" row-count="2" column-count="1" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
     <margin top="0" left="0" bottom="0" right="0"/>
     <constraints>
