+++ src/main/java/dev/chatcodes/ChatToolWindowFactory.java
@@ -1,16 +1,15 @@
-package com.jitterted;
+package dev.chatcodes;
 
 import com.intellij.openapi.project.DumbAware;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.wm.ToolWindow;
-import com.intellij.openapi.wm.ToolWindowFactory;
 import com.intellij.ui.content.Content;
 import com.intellij.ui.content.ContentFactory;
 import org.jetbrains.annotations.NotNull;
 
 import java.io.IOException;
 
-public class JitterChatToolWindowFactory implements ToolWindowFactory, DumbAware {
+public class ChatToolWindowFactory implements com.intellij.openapi.wm.ToolWindowFactory, DumbAware {
     // Create the tool window content.
     public void createToolWindowContent(Project project, ToolWindow toolWindow) {
         ChatToolWindow chatToolWindow = null;
