+++ src/main/resources/log4j2.properties
@@ -8,7 +8,7 @@ appender.xyz.layout.type = PatternLayout
 appender.xyz.layout.pattern = [%d{yy-MMM-dd HH:mm:ss:SSS}] [%p] [%c{1}:%L] - %m%n
 appender.xyz.fileName = /tmp/chatcodes.log
 
-rootLogger.level = debug
+rootLogger.level = info
 
 #rootLogger.appenderReferences
 rootLogger.appenderRefs = xyz
